Toque cuestiones esteticas menores.
El programa solo corre si la version de GSL 1.16 o mayor. De otro
modo el algoritmo de ajuste no lineal va a tirar un segmentation
fault

+++ idea-3.0/aux_functions.h
@@ -52,7 +52,7 @@ typedef struct peak_data
     int start_spr;
     int gamma;
     int start_gam;
-    double treshold;
+    float treshold;
     float * intensity;
     double ** bg;
     double *** intens;
@@ -128,7 +128,7 @@ void reset_all_seeds(gsl_vector * y, double ** seeds, int seeds_size, int n_peak
 
 void check(gsl_vector * y, double ** seeds, int seeds_size, int n_peaks, double ** bg, int bg_size);
 
-int check_for_null_peaks(double treshold, int numrings, int * zero_peak_index, float * intens);
+int check_for_null_peaks(float treshold, int numrings, int * zero_peak_index, float * intens);
 
 void set_seeds(int size, int * zero_peak_index, int exists, double ** seeds, double ** peak_seeds);
